Fix event bug causing an error in Firefox.
Remove superfluous event.preventDefault() from onSignUpFormSubmit() in
the SignUpFormContainer. The extra event action causes an error when the
user attempts to sign up in Firefox because the browser does not hold a
global symbol for event like Chrome and IE do. Anyway, the event's
defaults are already adequately handled by handleSubmit() in the
SignUpForm component.